requirements_pipeline.txt
pandas>=2.2
numpy>=1.26
yfinance>=0.2.36
requests>=2.31
python-dotenv>=1.0
